+++ NEWS.md
@@ -6,6 +6,7 @@ Version 1.4.9
   * Does not allow for a single random intercept
   * Does not yet allow for between random effect correlation
 * New formula interface for phylogenetic model adapted to trait model too
+* New phyplot.gllvm function for plotting the phylogenetic random effects
 
 Version 1.4.8
 =============